+++ jib-build-plan/kokoro/release_build.sh
@@ -1,0 +1,9 @@
+#!/bin/bash
+
+# Fail on any error.
+set -o errexit
+# Display commands to stderr.
+set -o xtrace
+
+cd github/jib
+./gradlew :jib-build-plan:prepareRelease