+++ jib-core/src/main/java/com/google/cloud/tools/jib/registry/credentials/DockerCredentialHelper.java
@@ -30,6 +30,11 @@ import java.io.InputStreamReader;
 import java.io.OutputStream;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
+import java.util.Arrays;
+import java.util.List;
+import java.util.Locale;
+import java.util.Properties;
+import java.util.function.Function;
 import javax.annotation.Nullable;
 
 /**
@@ -42,6 +47,8 @@ public class DockerCredentialHelper {
 
   private final String serverUrl;
   private final Path credentialHelper;
+  private final Properties systemProperties;
+  private Function<List<String>, ProcessBuilder> processBuilderFactory;
 
   /** Template for a Docker credential helper output. */
   @VisibleForTesting
@@ -66,8 +73,19 @@ public class DockerCredentialHelper {
    * @param credentialHelper the path to the credential helper executable
    */
   public DockerCredentialHelper(String serverUrl, Path credentialHelper) {
+    this(serverUrl, credentialHelper, System.getProperties(), ProcessBuilder::new);
+  }
+
+  @VisibleForTesting
+  DockerCredentialHelper(
+      String serverUrl,
+      Path credentialHelper,
+      Properties systemProperties,
+      Function<List<String>, ProcessBuilder> processBuilderFactory) {
     this.serverUrl = serverUrl;
     this.credentialHelper = credentialHelper;
+    this.systemProperties = systemProperties;
+    this.processBuilderFactory = processBuilderFactory;
   }
 
   /**
@@ -88,10 +106,31 @@ public class DockerCredentialHelper {
   public Credential retrieve()
       throws IOException, CredentialHelperUnhandledServerUrlException,
           CredentialHelperNotFoundException {
-    try {
-      String[] credentialHelperCommand = {credentialHelper.toString(), "get"};
+    boolean isWindows =
+        systemProperties.getProperty("os.name").toLowerCase(Locale.ENGLISH).contains("windows");
+    String lowerCaseHelper = credentialHelper.toString().toLowerCase(Locale.ENGLISH);
+    if (!isWindows || lowerCaseHelper.endsWith(".cmd") || lowerCaseHelper.endsWith(".exe")) {
+      return retrieve(Arrays.asList(credentialHelper.toString(), "get"));
+    }
 
-      Process process = new ProcessBuilder(credentialHelperCommand).start();
+    try {
+      return retrieve(Arrays.asList(credentialHelper.toString() + ".cmd", "get"));
+
+    } catch (CredentialHelperNotFoundException ex) {
+      try {
+        return retrieve(Arrays.asList(credentialHelper.toString(), "get"));
+      } catch (CredentialHelperNotFoundException ignored) {
+        throw ex;
+      }
+    }
+  }
+
+  private Credential retrieve(List<String> credentialHelperCommand)
+      throws IOException, CredentialHelperUnhandledServerUrlException,
+          CredentialHelperNotFoundException {
+    try {
+      Process process = processBuilderFactory.apply(credentialHelperCommand).start();
+
       try (OutputStream processStdin = process.getOutputStream()) {
         processStdin.write(serverUrl.getBytes(StandardCharsets.UTF_8));
       }