+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/GradleProjectProperties.java
@@ -16,25 +16,32 @@
 
 package com.google.cloud.tools.jib.gradle;
 
+import com.google.api.client.util.Lists;
 import com.google.cloud.tools.jib.api.Containerizer;
+import com.google.cloud.tools.jib.api.ImageReference;
+import com.google.cloud.tools.jib.api.InvalidImageReferenceException;
 import com.google.cloud.tools.jib.api.JavaContainerBuilder;
 import com.google.cloud.tools.jib.api.JibContainerBuilder;
 import com.google.cloud.tools.jib.api.LogEvent;
+import com.google.cloud.tools.jib.api.buildplan.ContainerBuildPlan;
 import com.google.cloud.tools.jib.event.events.ProgressEvent;
 import com.google.cloud.tools.jib.event.events.TimerEvent;
 import com.google.cloud.tools.jib.event.progress.ProgressEventHandler;
 import com.google.cloud.tools.jib.filesystem.DirectoryWalker;
 import com.google.cloud.tools.jib.filesystem.TempDirectoryProvider;
+import com.google.cloud.tools.jib.gradle.extension.JibGradlePluginExtension;
 import com.google.cloud.tools.jib.plugins.common.ContainerizingMode;
 import com.google.cloud.tools.jib.plugins.common.JavaContainerBuilderHelper;
 import com.google.cloud.tools.jib.plugins.common.ProjectProperties;
 import com.google.cloud.tools.jib.plugins.common.PropertyNames;
+import com.google.cloud.tools.jib.plugins.common.RawConfiguration.ExtensionConfiguration;
 import com.google.cloud.tools.jib.plugins.common.TimerEventHandler;
 import com.google.cloud.tools.jib.plugins.common.ZipUtil;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLogger;
 import com.google.cloud.tools.jib.plugins.common.logging.ConsoleLoggerBuilder;
 import com.google.cloud.tools.jib.plugins.common.logging.ProgressDisplayGenerator;
 import com.google.cloud.tools.jib.plugins.common.logging.SingleThreadedExecutor;
+import com.google.cloud.tools.jib.plugins.extension.JibPluginExtensionException;
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.base.Verify;
 import java.io.File;
@@ -45,6 +52,8 @@ import java.nio.file.Paths;
 import java.time.Duration;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.ServiceLoader;
+import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import javax.annotation.Nullable;
 import org.apache.tools.ant.taskdefs.condition.Os;
@@ -314,7 +323,7 @@ public class GradleProjectProperties implements ProjectProperties {
 
   @Nullable
   @Override
-  public String getMainClassFromJar() {
+  public String getMainClassFromJarPlugin() {
     Jar jarTask = (Jar) project.getTasks().findByName("jar");
     if (jarTask == null) {
       return null;
@@ -389,7 +398,61 @@ public class GradleProjectProperties implements ProjectProperties {
   }
 
   @Override
-  public JibContainerBuilder runPluginExtensions(JibContainerBuilder jibContainerBuilder) {
-    return jibContainerBuilder;
+  public JibContainerBuilder runPluginExtensions(
+      List<? extends ExtensionConfiguration> extensionConfigs,
+      JibContainerBuilder jibContainerBuilder)
+      throws JibPluginExtensionException {
+    List<JibGradlePluginExtension> services =
+        Lists.newArrayList(ServiceLoader.load(JibGradlePluginExtension.class).iterator());
+    return runPluginExtensions(services, extensionConfigs, jibContainerBuilder);
+  }
+
+  @VisibleForTesting
+  JibContainerBuilder runPluginExtensions(
+      List<JibGradlePluginExtension> services,
+      List<? extends ExtensionConfiguration> extensionConfigs,
+      JibContainerBuilder jibContainerBuilder)
+      throws JibPluginExtensionException {
+    if (extensionConfigs.isEmpty()) {
+      log(LogEvent.debug("No Jib plugin extensions configured to load"));
+      return jibContainerBuilder;
+    }
+
+    JibGradlePluginExtension extension = null;
+    ContainerBuildPlan buildPlan = jibContainerBuilder.toContainerBuildPlan();
+    GradleExtensionLogger extensionLogger = new GradleExtensionLogger(this::log);
+    try {
+      for (ExtensionConfiguration config : extensionConfigs) {
+        String extensionClass = config.getExtensionClass();
+        extension = findConfiguredExtension(services, extensionClass);
+        if (extension == null) {
+          throw new JibPluginExtensionException(
+              JibGradlePluginExtension.class,
+              "extension configured but not discovered on Jib runtime classpath: "
+                  + extensionClass);
+        }
+
+        log(LogEvent.lifecycle("Running extension: " + extensionClass));
+        buildPlan =
+            extension.extendContainerBuildPlan(
+                buildPlan, config.getProperties(), () -> project, extensionLogger);
+        ImageReference.parse(buildPlan.getBaseImage()); // to validate image reference
+      }
+      return jibContainerBuilder.applyContainerBuildPlan(buildPlan);
+
+    } catch (InvalidImageReferenceException ex) {
+      throw new JibPluginExtensionException(
+          Verify.verifyNotNull(extension).getClass(),
+          "invalid base image reference: " + buildPlan.getBaseImage(),
+          ex);
+    }
+  }
+
+  @Nullable
+  private JibGradlePluginExtension findConfiguredExtension(
+      List<JibGradlePluginExtension> extensions, String extensionClass) {
+    Predicate<JibGradlePluginExtension> matchesClassName =
+        extension -> extension.getClass().getName().equals(extensionClass);
+    return extensions.stream().filter(matchesClassName).findFirst().orElse(null);
   }
 }