+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenRawConfiguration.java
@@ -16,11 +16,14 @@
 
 package com.google.cloud.tools.jib.maven;
 
+import com.google.cloud.tools.jib.api.buildplan.AbsoluteUnixPath;
 import com.google.cloud.tools.jib.api.buildplan.FilePermissions;
 import com.google.cloud.tools.jib.api.buildplan.ImageFormat;
+import com.google.cloud.tools.jib.maven.JibPluginConfiguration.ExtraDirectoryParameters;
 import com.google.cloud.tools.jib.plugins.common.AuthProperty;
 import com.google.cloud.tools.jib.plugins.common.RawConfiguration;
 import java.nio.file.Path;
+import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -161,8 +164,13 @@ public class MavenRawConfiguration implements RawConfiguration {
   }
 
   @Override
-  public List<Path> getExtraDirectories() {
-    return MojoCommon.getExtraDirectories(jibPluginConfiguration);
+  public Map<Path, AbsoluteUnixPath> getExtraDirectories() {
+    Map<Path, AbsoluteUnixPath> directoryMap = new LinkedHashMap<>();
+    for (ExtraDirectoryParameters extraDirectory :
+        MojoCommon.getExtraDirectories(jibPluginConfiguration)) {
+      directoryMap.put(extraDirectory.getFrom(), AbsoluteUnixPath.get(extraDirectory.getInto()));
+    }
+    return directoryMap;
   }
 
   @Override
@@ -203,5 +211,10 @@ public class MavenRawConfiguration implements RawConfiguration {
   @Override
   public Path getImageJsonOutputPath() {
     return jibPluginConfiguration.getImageJsonOutputPath();
+  }
+
+  @Override
+  public List<? extends ExtensionConfiguration> getPluginExtensions() {
+    return jibPluginConfiguration.getPluginExtensions();
   }
 }