+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/PluginConfigurationProcessor.java
@@ -52,6 +52,7 @@ import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
 import java.util.HashSet;
 import java.util.List;
+import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
 import java.util.function.BiFunction;
@@ -125,7 +126,9 @@ public class PluginConfigurationProcessor {
         processCommonConfiguration(
             rawConfiguration, inferredAuthProvider, projectProperties, containerizer);
     JibContainerBuilder updatedContainerBuilder =
-        projectProperties.runPluginExtensions(jibContainerBuilder).setFormat(ImageFormat.Docker);
+        projectProperties
+            .runPluginExtensions(rawConfiguration.getPluginExtensions(), jibContainerBuilder)
+            .setFormat(ImageFormat.Docker);
 
     return JibBuildRunner.forBuildToDockerDaemon(
             updatedContainerBuilder,
@@ -184,7 +187,8 @@ public class PluginConfigurationProcessor {
         processCommonConfiguration(
             rawConfiguration, inferredAuthProvider, projectProperties, containerizer);
     JibContainerBuilder updatedContainerBuilder =
-        projectProperties.runPluginExtensions(jibContainerBuilder);
+        projectProperties.runPluginExtensions(
+            rawConfiguration.getPluginExtensions(), jibContainerBuilder);
 
     return JibBuildRunner.forBuildTar(
             updatedContainerBuilder,
@@ -258,7 +262,8 @@ public class PluginConfigurationProcessor {
         processCommonConfiguration(
             rawConfiguration, inferredAuthProvider, projectProperties, containerizer);
     JibContainerBuilder updatedContainerBuilder =
-        projectProperties.runPluginExtensions(jibContainerBuilder);
+        projectProperties.runPluginExtensions(
+            rawConfiguration.getPluginExtensions(), jibContainerBuilder);
 
     return JibBuildRunner.forBuildImage(
             updatedContainerBuilder,
@@ -392,11 +397,15 @@ public class PluginConfigurationProcessor {
         getCreationTime(rawConfiguration.getCreationTime(), projectProperties));
 
     // Adds all the extra files.
-    for (Path directory : rawConfiguration.getExtraDirectories()) {
-      if (Files.exists(directory)) {
+    for (Map.Entry<Path, AbsoluteUnixPath> entry :
+        rawConfiguration.getExtraDirectories().entrySet()) {
+      Path sourceDirectory = entry.getKey();
+      AbsoluteUnixPath targetDirectory = entry.getValue();
+      if (Files.exists(sourceDirectory)) {
         jibContainerBuilder.addFileEntriesLayer(
             JavaContainerBuilderHelper.extraDirectoryLayerConfiguration(
-                directory,
+                sourceDirectory,
+                targetDirectory,
                 rawConfiguration.getExtraDirectoryPermissions(),
                 modificationTimeProvider));
       }